property_management/shared/requirements.txt
# FastAPI и связанные зависимости
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Кэширование
redis==5.0.1

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP клиент
httpx==0.25.2

# Логирование
structlog==23.2.0

# Валидация и утилиты
email-validator==2.1.0
python-dateutil==2.8.2

# Мониторинг
prometheus-client==0.19.0
